requirements.txt
streamlit

python-dotenv

pydantic

sentence-transformers

faiss-cpu

langchain
langchain-community
langchain-groq

PyPDF2
